fix: Prevent absolute path creation in uv lock template (#2769)
This change fixes a bug in the `lock` rule where, when the package is at
the root level, the path to `requirements.txt` is constructed
incorrectly with a leading double slash (`//requirements.txt`), causing
it to be interpreted as an absolute path.

This change detects if the package is empty before constructing the
output path.

Work towards #1975

---------

Co-authored-by: Ignas Anikevicius <240938+aignas@users.noreply.github.com>