fix: declare PyInfo as provided by test/binary/library (#2777)
Currently, the rules don't advertise the PyInfo provider through the
provides argument
to the rule function. This means that aspects that want to consume
PyInfo can't use
`required_providers` to restrict themselves to the Python rules, and
instead have to
apply to all rules.

To fix, add PyInfo to the provides arg of the rules.

Fixes https://github.com/bazel-contrib/rules_python/issues/2506

---------

Co-authored-by: Richard Levasseur <rlevasseur@google.com>
Co-authored-by: Richard Levasseur <richardlev@gmail.com>