feat: Have `pip_compile` generate a `*.test` target; deprecate `*_test` (#2812)
Fixes #2794.

The `pip_compile` macro generates `*_test` and `*.update` targets. This
pattern does not match with other macros that generate similar targets,
namely `gazelle_python_manifest` and uv `lock` (though that's `.run`
instead of `.test` but either way, it uses a dot `.` instead of
underscore
`_`).

Adjust the macro so that a `.test` target is made. The `_test` target is
aliased with a deprecation warning, to be removed in the next major
version.