fix(pypi): call python --version before marker eval (#2819)
`bzlmod` has the full python version information statically and we don't
need to call Python to get its version, but for `WORKSPACE` that is not
the case and we have to call it before evaluating the markers in
universal requirements files.

This also fixes transitions in the `compile_pip_requirements` macro
where
the `.update` target would not transition correctly based on the
`python_version` parameter.

Fixes #2818