fix: requires_file preserves extras that package depends on (#2807)
When requirements are passed in through `requires_file` the extras are
not preserved.

eg if the contents of requires file is `example[extras]==1.1.1`, bazel
will currently write to the METADATA file `Requires-Dist:
example==1.1.1`. This PR attempts to fix that by adding that back if
there are any extras.

The expected output should be `Requires-Dist: example[extras]==1.1.1`