refactor: make env marker config available through target and flag (#2853)
This factors creation of (most of) the env marker dict into a separate
target and
provides a label flag to allow customizing the target that provides it.

This makes it easier for users to override how env marker values are
computed. The
`env_marker_setting` rule will still, if necessary, compute values from
the toolchain,
but existing keys (computed from the env marker config target) have
precedence.

The `EnvMarkerInfo` provider is the interface for implementing a custom
env marker
config target; it will be publically exposed in a subsequent PR.

Along the way, unify how the env dict and defaults are set.

Work towards https://github.com/bazel-contrib/rules_python/issues/2826