refactor: have bzlmod pass platforms to python_register_toolchains (#2884)
This is to facilitate eventually allowing overrides to add additional
platforms.

Instead of the PLATFORMS global being used, the python bzlmod extension
passing the
mapping directly to python_register_toolchains, then receives back the
subset of
platforms that had repos defined. That subset is then later used when
(re)constructing
the list of repo names for the toolchains.