sphinxdocs: make xrefs to tag class attributes using attr role work (#2895)
The role assigned to attributes within the tag class directive were
being given the
role `arg`, when they should be `attr`. This caused xrefs using the attr
role to
be unable to find them.

To fix, set them to have the correct role, like the repo rule and
regular rule
directives do. Also add a test.