refactor: make bzlmod directly aware of created toolchain repo names (#2885)
This makes python_register_toolchains return the repo names it created,
which allows
the bzlmod code to be directly aware of the repos that were created
instead of having
to rely on assuming the names via the platform keys.

This is to facilitate python_register_toolchains creating a more
arbitrary subset
of platform-specific repos.

Work towards https://github.com/bazel-contrib/rules_python/issues/2081