fix: Normalize main script path in Python bootstrap (#2925)
Use `os.path.normpath()` to resolve `_main/../repo/` to `repo/` and
convert forward
slashes to backward slashes on Windows.

This fixes an issue where `_main` doesn't exist within runfiles and in
turn the
later assertion that the path to main exists fails (~L542). This
happens,
for example, when packaging a `py_binary` from a foreign repo into a
tar/container.

---------

Co-authored-by: Richard Levasseur <richardlev@gmail.com>
Co-authored-by: Richard Levasseur <rlevasseur@google.com>