refactor: change site_packages_symlinks to venv_symlinks (#2939)
This generalizes the ability to populate the venv directory by adding
and additional field,
`kind`, which tells which directory of the venv to populate. A symbolic
constant is used
to indicate which directory so that users don't have to re-derive the
platform and version
specific paths that make up the venv directory names.

This follows the design described by
https://github.com/bazel-contrib/rules_python/issues/2156#issuecomment-2855580026

This also changes it to a depset of structs to make it more forward
compatible. A provider
is used because they're slightly more memory efficient than regular
structs.

Work towards https://github.com/bazel-contrib/rules_python/issues/2156