docs: fix sphinxdocs mis-redirect (#2940)
The redirect was going to a non-existent URL when viewed on the deployed
docs.

This was happening because the absolute paths `/api/whatever` don't
exist in the deployed
site -- it's actually `/en/latest/api/whatever`. This went unnoticed
because it works
locally (where there is no /en/latest prefix).

To fix, use a relative url (relative urls are relative to the path that
is redirected from)