feat: allow custom platform when overriding (#2880)
This basically allows using any python-build-standalone archive and
using it
if custom flags are set. This is done through the
`single_version_platform_override()`
API, because such archives are inherently version and platform specific.

Key changes:
* The `platform` arg can be any value (mostly; it ends up in repo names)
* Added `target_compatible_with` and `target_settings` args, which
become the
  settings used on the generated toolchain() definition.

The platform settings are version specific, i.e. the key
`(python_version, platform)`
is what maps to the TCW/TS values.

If an existing platform is used, it'll override the defaults that
normally come
from the PLATFORMS global for the particular version. If a new platform
is used,
it creates a new platform entry with those settings.

Along the way:
* Added various docs about internal variables so they're easier to grok
at a glance.

Work towards https://github.com/bazel-contrib/rules_python/issues/2081