revert: change default bootstrap back to system_python (#2968)
Switch the default bootstrap back to system_python, per maintainer
discussion.

The main reason is downstream consumers are unlikely to be fully ready
for the
usage of raw symlinks (declare_symlink artifacts). APIs to detect them
aren't
available until Bazel 8, which makes it difficult for packaging rules,
such
as rules_pkg, bazel-lib, or tar rules.

This reverts the core part of commit
9f3512fe0cc6d7229170e45724e22e64be0b8300