fix: use platform_info.target_settings in toolchain aliases (#3001)
During the refactor we forgot one more place where the `flag_values` on
the platform information was used. They were no longer populated and
broke.

The solution is to use `selects.config_setting_group` to maintain
behaviour and in order to smoke test I have added a target to verify
that the aliases work.

Related to #2875
Fixes #2993

Co-authored-by: Richard Levasseur <richardlev@gmail.com>