feat(gazelle): Add type-checking only dependencies to pyi_deps (#3014)
https://github.com/bazel-contrib/rules_python/pull/2538 added the
attribute `pyi_deps` to python rules, intended to be used for
dependencies that are only used for type-checking purposes. This PR adds
a new directive, `gazelle:python_generate_pyi_deps`, which, when
enabled:

- When a dependency is added only to satisfy type-checking only imports
(in a `if TYPE_CHECKING:` block), the dependency is added to `pyi_deps`
instead of `deps`;
- Third-party stub packages (eg. `boto3-stubs`) are now added to
`pyi_deps` instead of `deps`.

---------

Co-authored-by: Douglas Thor <dougthor42@users.noreply.github.com>