fix(gazelle) Fix dependency added as both deps and pyi_deps (#3036)
Fix an issue in https://github.com/bazel-contrib/rules_python/pull/3014
where a dependency may end up being added in both `deps` and `pyi_deps`,
in cases where the regular and the type-checking import refer to
different python modules on the same `py_library` target. Other cases
are already deduplicated earlier on, but this case can only be
deduplicated in the resolve phase.

(No new changelog entry since this is a fix to an unreleased feature
that is already in the changelog)