fix(pypi): only generate namespace package shims if implicit namespaces are disabled (#3059)
The refactoring to move the pkgutil shim generation to build phase
inverted
the logic for when it should be activated. When
`enable_implicit_namespace_pkgs=True`, it means to not generate the
pkgutil shims
("respect the Python definition of the namespace package").

To fix, just invert the logic that activates it.

A test will be added in a subsequent PR because the necessary helper
isn't
in the 1.5 branch.

Fixes https://github.com/bazel-contrib/rules_python/issues/3038

---------

Co-authored-by: Ignas Anikevicius <240938+aignas@users.noreply.github.com>