fix(toolchains): fix the URLs and sha256 values (#3070)
It seems that in #3062 we did not notice that the latest builds for
`linux-aarch64-freethreaded` actually needed a different build tag and
this is the case only for the latest releases.

What is more, the `windows-aarch64` build had a wrong strip-prefix.

This is fixing all of these issues. #3028 will make the 404 error
messages from `curl` more visible, so the first issue will be more
easily caught during authoring, whereas the second issue is going to be
more likely caught via code review, or a CI that is exercising the
actual toolchains.